changed code following clean code guidelines
