+++ src/App.tsx
@@ -1,26 +1,30 @@
 import { Routes, Route, BrowserRouter } from "react-router-dom";
-import { observer } from "mobx-react";
-
 import Login from "./components/Login";
 
 import Accounts from "./components/Accounts";
 
-import Profile from "./components/Profile";
+import Profile from "./components/Profile/Profile";
 import { TransactionsStoreContextProvider } from "./context/StoresContext";
 
-import TransactionRoute from "./components/TranscationRoute";
+import TransactionRoute from "./components/TranscationRoute/TransactionRoute";
 
 import "./App.css";
+import {
+  HOME_PATH,
+  LOGIN_IN_PATH,
+  PROFILE_PATH,
+  TRANSACTIONS_PATH,
+} from "./constants/PathConstants";
 
 const App = () => {
   return (
     <TransactionsStoreContextProvider>
       <BrowserRouter basename={process.env.PUBLIC_URL}>
         <Routes>
-          <Route path="/login" element={<Login />} />
-          <Route path="/" element={<Accounts />} />
-          <Route path="/transactions" element={<TransactionRoute />} />
-          <Route path="/profile" element={<Profile />} />
+          <Route path={LOGIN_IN_PATH} element={<Login />} />
+          <Route path={HOME_PATH} element={<Accounts />} />
+          <Route path={TRANSACTIONS_PATH} element={<TransactionRoute />} />
+          <Route path={PROFILE_PATH} element={<Profile />} />
         </Routes>
       </BrowserRouter>
     </TransactionsStoreContextProvider>
