+++ src/components/DebitBox/index.css
@@ -6,9 +6,10 @@
     padding: 20px;
     padding-left: 30px;
     padding-right: 30px;
-    padding-bottom: 0;
+    padding-bottom: 20px;
     display: flex;
     flex-grow: 1;
+    align-items: center;
     justify-content: space-around;
 }
