+++ src/components/TranscationsCard/index.tsx
@@ -1,14 +1,3 @@
-import { observer } from "mobx-react";
-import "./index.css";
-import TxnList from "../TxnList";
+import TransactionsCard from "./TransactionsCard";
 
-const TransactionsCard = () => {
-  return (
-    <div className="last-txn-card">
-      <h1>Last Transactions</h1>
-      <TxnList />
-    </div>
-  );
-};
-
-export default observer(TransactionsCard);
+export default TransactionsCard;
