+++ src/constants/PathConstants.ts
@@ -1,0 +1,4 @@
+export const LOGIN_IN_PATH = '/login'
+export const HOME_PATH = '/'
+export const TRANSACTIONS_PATH = '/transactions'
+export const PROFILE_PATH = '/profile'
