--- src/constants/apistatus.ts
@@ -1,5 +1,0 @@
-
-type Status = "INITIAL" | "SUCCESS" | "FAILED" | "LOADING"
-
-
-export default Status;
