+++ src/context/StoresContext.tsx
@@ -1,4 +1,4 @@
-import { PropsWithChildren, createContext, useContext, useRef } from "react";
+import { PropsWithChildren, createContext, useRef } from "react";
 import TransactionsStore from "../stores/TransactionsStore";
 
 export const TransactionStoreContext = createContext<
