+++ src/hooks/FetchUserId/UseUserId.tsx
@@ -1,5 +1,5 @@
 import Cookies from "js-cookie";
-import Details from "../../types/detailstype";
+import { Details } from "../../types/detailstype";
 const useUserId = (): Details => {
   const newUserCreds = Cookies.get("secret_token");
   return newUserCreds === undefined ? undefined : JSON.parse(newUserCreds);
