+++ src/stores/models/TransactionObjectmodel.ts
@@ -13,9 +13,8 @@ class TransactionModel {
 
     constructor(transaction: TransactionObj){
         makeObservable(this, {
-            id: observable,
             transactionName: observable,
-            type:observable ,amount : observable  ,userId : observable   ,date : observable    ,category : observable,
+            type:observable ,amount : observable  ,date : observable    ,category : observable,
             setType: action.bound,
             setAmount: action.bound,
             setTransactionName: action.bound,
