+++ src/types/detailstype.ts
@@ -1,10 +1,8 @@
-
-type Details = 
+export type Status = "INITIAL" | "SUCCESS" | "FAILED" | "LOADING"
+export type Details = 
   {
     secretToken: string;
     username: string | undefined;
     userId: number | undefined;
     isAdmin: boolean;
 } | undefined
-
-export default Details
