+++ src/utils/recentTxnsDateConverter.ts
@@ -1,23 +1,22 @@
-const WeekDays = ["Sun", "Mon", "Tue", "Wed", "Thu", "Fri", "Sat"];
+const weekDays = ["Sun", "Mon", "Tue", "Wed", "Thu", "Fri", "Sat"];
 
 type DataArray = {
-  date: Date,
-  sum: number,
-  type: string
-}
+  date: Date;
+  sum: number;
+  type: string;
+};
 
 type Data = {
-  last_7_days_transactions_credit_debit_totals: DataArray []
-}
+  last_7_days_transactions_credit_debit_totals: DataArray[];
+};
 
 type Days = {
-    day: string;
-    debit: number;
-    credit: number;
-}[]
+  day: string;
+  debit: number;
+  credit: number;
+}[];
 
-
-const GetReqFormat = (sevenDaysTxns: Data) => {
+const recentTxnsDateConverter = (sevenDaysTxns: Data) => {
   const day0 = new Date();
   const day1 = new Date();
   day1.setDate(new Date().getDate() - 1);
@@ -33,33 +32,33 @@ const GetReqFormat = (sevenDaysTxns: Data) => {
   day6.setDate(new Date().getDate() - 6);
 
   let sevenWeeksData: Days = [
-    { day: WeekDays[day6.getDay()], debit: 0, credit: 0 },
-    { day: WeekDays[day5.getDay()], debit: 0, credit: 0 },
-    { day: WeekDays[day4.getDay()], debit: 0, credit: 0 },
-    { day: WeekDays[day3.getDay()], debit: 0, credit: 0 },
-    { day: WeekDays[day2.getDay()], debit: 0, credit: 0 },
-    { day: WeekDays[day1.getDay()], debit: 0, credit: 0 },
-    { day: WeekDays[day0.getDay()], debit: 0, credit: 0 },
+    { day: weekDays[day6.getDay()], debit: 0, credit: 0 },
+    { day: weekDays[day5.getDay()], debit: 0, credit: 0 },
+    { day: weekDays[day4.getDay()], debit: 0, credit: 0 },
+    { day: weekDays[day3.getDay()], debit: 0, credit: 0 },
+    { day: weekDays[day2.getDay()], debit: 0, credit: 0 },
+    { day: weekDays[day1.getDay()], debit: 0, credit: 0 },
+    { day: weekDays[day0.getDay()], debit: 0, credit: 0 },
   ];
 
+  const givenData = sevenDaysTxns.last_7_days_transactions_credit_debit_totals;
 
- const GivenData = sevenDaysTxns.last_7_days_transactions_credit_debit_totals
-
-for (let item of GivenData){
-  const dateOfTxn = new Date(item.date);
- const dayOfTxn = WeekDays[dateOfTxn.getDay()];
-  const modified = sevenWeeksData.map((eachData) => {
-    if (eachData.day === dayOfTxn) {
-        if (item.type === "credit"){
-          eachData["credit"] += item.sum} else{
-            eachData["debit"] += item.sum
+  for (let item of givenData) {
+    const dateOfTxn = new Date(item.date);
+    const dayOfTxn = weekDays[dateOfTxn.getDay()];
+    const modified = sevenWeeksData.map((eachData) => {
+      if (eachData.day === dayOfTxn) {
+        if (item.type === "credit") {
+          eachData["credit"] += item.sum;
+        } else {
+          eachData["debit"] += item.sum;
         }
-    }
-    return eachData
-  })
-  sevenWeeksData = modified
-}
+      }
+      return eachData;
+    });
+    sevenWeeksData = modified;
+  }
   return sevenWeeksData;
 };
 
-export default GetReqFormat;
+export default recentTxnsDateConverter;
